Remove Experiment class.
This is a major refactor of the code, removing the Experiment class as I
am using wandb instead for experiment management.
As this made the scripts in experiments/scripts obsolete, I have removed
these.

Using semble, code duplication has been reduced in the data generation
scripts.
Additionally, the new dictionary-based description of the trajectory
sampler will allow for better backwards-compatibility.